feat: machine keys
