+++ coderd/database/models.go
@@ -120,6 +120,7 @@ const (
 	LoginTypePassword LoginType = "password"
 	LoginTypeGithub   LoginType = "github"
 	LoginTypeOIDC     LoginType = "oidc"
+	LoginTypeMachine  LoginType = "machine"
 )
 
 func (e *LoginType) Scan(src interface{}) error {
